requirements.txt
Pillow==8.2.0
shutil==1.7.9
